--- a/dpctl/tensor/libtensor/source/linalg_functions/dot_dispatch.hpp
+++ b/dpctl/tensor/libtensor/source/linalg_functions/dot_dispatch.hpp
@@ -60,7 +60,6 @@ template <typename T1, typename T2> struct DotNoAtomicOutputType
 {
     using value_type = typename std::disjunction< // disjunction is C++17
                                                   // feature, supported by DPC++
-        td_ns::BinaryTypeMapResultEntry<T1, bool, T2, bool, bool>,
         td_ns::BinaryTypeMapResultEntry<T1, bool, T2, bool, std::uint8_t>,
         td_ns::BinaryTypeMapResultEntry<T1,
                                         std::uint8_t,