--- a/dpctl/tests/test_usm_ndarray_linalg.py
+++ b/dpctl/tests/test_usm_ndarray_linalg.py
@@ -571,6 +571,41 @@ def test_matmul_inplace_same_tensors():
     assert dpt.all(ar2 == dpt.full(sh, n, dtype=ar2.dtype))
 
 
+@pytest.fixture
+def random_matrix():
+    rs = np.random.RandomState(seed=123456)
+    m_np = rs.randint(low=0, high=6, size=(400, 400))
+    return m_np
+
+
+@pytest.mark.parametrize("dtype", _numeric_types)
+def test_matmul_largish(dtype, random_matrix):
+    q = get_queue_or_skip()
+    skip_if_dtype_not_supported(dtype, q)
+
+    m_np = random_matrix.astype(dtype)
+    x_np = np.matmul(m_np.T, m_np)
+
+    m = dpt.asarray(m_np)
+    mT = dpt.asarray(m.mT, copy=True, order="C")
+    x1 = dpt.matmul(m.mT, m)
+    x2 = dpt.matmul(mT, m)
+
+    assert dpt.allclose(x1, x2)
+    assert dpt.allclose(x1, dpt.asarray(x_np))
+
+    m_np = m_np[:-1, :-1]
+    x_np = np.matmul(m_np.T, m_np)
+
+    m = m[:-1, :-1]
+    mT = dpt.asarray(m.mT, copy=True, order="C")
+    x1 = dpt.matmul(m.mT, m)
+    x2 = dpt.matmul(mT, m)
+
+    assert dpt.allclose(x1, x2)
+    assert dpt.allclose(x1, dpt.asarray(x_np))
+
+
 @pytest.mark.parametrize("dtype", _numeric_types)
 def test_tensordot_outer(dtype):
     q = get_queue_or_skip()