Refactoring of gemm implementation, adding faster gemm kernel implementation.

This change gets rid of all non-batch functors, modularizes
duplicated code, and implement non-batches functions as calls
to batched functors with trivial constexpr batch indexer.

This change also adds faster gemm kernel that threads of N,M space,
and accumulates entire range of K in single work-item.

Dispatch logic changed too, we dispatch to thead-K kernel only if
(n,m) space is sufficiently small.

A test is added to exercise new kernel for all supported types.

- [x] Have you provided a meaningful PR description?
- [x] Have you added a test, reproducer or referred to an issue with a reproducer?
- [x] Have you tested your changes locally for CPU and GPU devices?
- [x] Have you made sure that new changes do not introduce compiler warnings?
- [x] Have you checked performance impact of proposed changes?
- [ ] If this PR is a work in progress, are you opening the PR as a draft?