@ndgrigorian 2024-02-10
The change to the complex hyperparameters only slightly impacted performance.

Before
```
In [3]: x1 = dpt.ones((2000, 2000), dtype="c8", device="cpu")

In [4]: x1 = dpt.ones((2000, 2000), dtype="c8")

In [5]: x2 = dpt.ones((2000, 2000), dtype="c8")

In [6]: %timeit z = dpt.matmul(x1, x2)
153 ms ± 17.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [7]: %timeit z = dpt.matmul(x1, x2)
145 ms ± 14.7 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

After
```
In [2]: x1 = dpt.ones((2000, 2000), dtype="c8")
x
In [3]: x2 = dpt.ones((2000, 2000), dtype="c8")

In [4]: %timeit z = dpt.matmul(x1, x2)
221 ms ± 3.59 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [5]: %timeit z = dpt.matmul(x1, x2)
213 ms ± 10.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```