@ndgrigorian 2024-02-10
Complex batched case appears to be broken.
Inflection point is at exactly 256

```
In [17]: dpt.matmul(x1, x2)
Out[17]:
usm_ndarray([[[256.+0.j, 256.+0.j, 256.+0.j, ..., 256.+0.j, 256.+0.j,
               256.+0.j],
              [256.+0.j, 256.+0.j, 256.+0.j, ..., 256.+0.j, 256.+0.j,
               256.+0.j],
              [256.+0.j, 256.+0.j, 256.+0.j, ..., 256.+0.j, 256.+0.j,
               256.+0.j],
              ...,
              [256.+0.j, 256.+0.j, 256.+0.j, ..., 256.+0.j, 256.+0.j,
               256.+0.j],
              [256.+0.j, 256.+0.j, 256.+0.j, ..., 256.+0.j, 256.+0.j,
               256.+0.j],
              [256.+0.j, 256.+0.j, 256.+0.j, ..., 256.+0.j, 256.+0.j,
               256.+0.j]],

             [[  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              ...,
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j]],

             [[  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              ...,
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j]],

             [[  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              ...,
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j]],

             [[  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              ...,
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j],
              [  0.+0.j,   0.+0.j,   0.+0.j, ...,   0.+0.j,   0.+0.j,
                 0.+0.j]]], dtype=complex64)

In [18]: x2 = dpt.ones((5, 255, 255), dtype="c8")

In [19]: x1 = dpt.ones((5, 255, 255), dtype="c8")

In [20]: dpt.matmul(x1, x2)
Out[20]:
usm_ndarray([[[255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              ...,
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j]],

             [[255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              ...,
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j]],

             [[255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              ...,
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j]],

             [[255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              ...,
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j]],

             [[255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              ...,
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j],
              [255.+0.j, 255.+0.j, 255.+0.j, ..., 255.+0.j, 255.+0.j,
               255.+0.j]]], dtype=complex64)
```