@oleksandr-pavlyk 2024-02-10
> Complex batched contiguous case appears to be broken. Inflection point is at exactly 256> 
> This appears to happen for int16, uint16, etc. as well. Does not impact strided case.

Thanks @ndgrigorian . This was caused by an oversight, and would impact all contiguous inputs that would normally dispatch to tree implementation. This is fixed now.